refactor: extract attachment imports from AgentWorkspace
